pyproject.toml
[project]
name = "sub-convert"
version = "0.4.5"
description = "Server-side subscription converter for Base64 v2rayN sources to Clash and sing-box outputs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0,<7",
    "requests>=2.32,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["subscription_converter*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
